Move agent_auxiliars script to utils/
